+++ src/main/kotlin/org/vitrivr/cottontail/cli/schema/DropSchemaCommand.kt
@@ -1,8 +1,8 @@
 package org.vitrivr.cottontail.cli.schema
 
-import com.github.ajalt.clikt.parameters.options.convert
+import com.github.ajalt.clikt.output.TermUi
+import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.options.prompt
 import org.vitrivr.cottontail.database.queries.binding.extensions.proto
 import org.vitrivr.cottontail.grpc.CottontailGrpc
 import org.vitrivr.cottontail.grpc.DDLGrpc
@@ -17,23 +17,23 @@ import kotlin.time.measureTimedValue
  * @version 1.0.1
  */
 @ExperimentalTime
-class DropSchemaCommand(private val ddlStub: DDLGrpc.DDLBlockingStub) : AbstractSchemaCommand(name = "drop", help = "Drops the schema with the given name. Usage: schema drop <name>") {
+class DropSchemaCommand(private val ddlStub: DDLGrpc.DDLBlockingStub) :
+    AbstractSchemaCommand(name = "drop", help = "Drops the schema with the given name. Usage: schema drop <name>") {
     /** Flag that can be used to directly provide confirmation. */
     private val confirm: Boolean by option(
         "-c",
         "--confirm",
         help = "Directly provides the confirmation option."
-    ).convert {
-        it.toLowerCase() == "y"
-    }.prompt(
-        "Do you really want to drop the schema ${this.schemaName} [y/N]?",
-        default = "n",
-        showDefault = false
-    )
+    ).flag()
 
 
     override fun exec() {
-        if (this.confirm) {
+        if (this.confirm || TermUi.confirm(
+                "Do you really want to drop the schema ${this.schemaName} [y/N]?",
+                default = false,
+                showDefault = false
+            ) == true
+        ) {
             /* Execute query. */
             val timedTable = measureTimedValue {
                 TabulationUtilities.tabulate(