+++ src/main/kotlin/org/vitrivr/cottontail/database/general/AbstractTx.kt
@@ -10,7 +10,7 @@ import org.vitrivr.cottontail.model.exceptions.TxException
  * An abstract [Tx] implementation that provides some basic functionality.
  *
  * @author Ralph Gasser
- * @version 1.3.0
+ * @version 1.4.0
  */
 abstract class AbstractTx(override val context: TransactionContext) : Tx {
     /** Flag indicating whether or not this [IndexTx] was closed */
@@ -26,9 +26,15 @@ abstract class AbstractTx(override val context: TransactionContext) : Tx {
      * actual commit.
      */
     final override fun commit() {
-        if (this.status == TxStatus.DIRTY) {
-            this.snapshot.commit()
-            this.status = TxStatus.CLEAN
+        try {
+            if (this.status == TxStatus.DIRTY) {
+                this.snapshot.commit()
+            } else if (this.status == TxStatus.ERROR) {
+                throw IllegalArgumentException("Transaction ${this.context.txId} cannot be committed, because it is is an error state.")
+            }
+        } finally {
+            this.status = TxStatus.CLOSED
+            this.cleanup()
         }
     }
 
@@ -40,21 +46,13 @@ abstract class AbstractTx(override val context: TransactionContext) : Tx {
      * commit.
      */
     final override fun rollback() {
-        if (this.status == TxStatus.DIRTY || this.status == TxStatus.ERROR) {
-            this.snapshot.rollback()
-            this.status = TxStatus.CLEAN
-        }
-    }
-
-    /**
-     * Closes this [AbstractTx]. If there are uncommitted changes, these changes will be rolled back.
-     * Closed [AbstractTx] cannot be used anymore!
-     */
-    final override fun close() {
-        if (this.status != TxStatus.CLOSED) {
-            this.rollback()
-            this.cleanup()
+        try {
+            if (this.status == TxStatus.DIRTY || this.status == TxStatus.ERROR) {
+                this.snapshot.rollback()
+            }
+        } finally {
             this.status = TxStatus.CLOSED
+            this.cleanup()
         }
     }