+++ src/test/kotlin/org/vitrivr/cottontail/TestConstants.kt
@@ -1,8 +1,8 @@
 package org.vitrivr.cottontail
 
 import org.vitrivr.cottontail.config.Config
-import org.vitrivr.cottontail.config.MapDBConfig
-import java.nio.file.Paths
+import java.nio.file.Files
+import java.util.*
 
 /**
  * Some constants used during execution of unit tests.
@@ -11,19 +11,8 @@ import java.nio.file.Paths
  * @version 1.1.0
  */
 object TestConstants {
-
-    init {
-    }
-
-    /** [Config] used for executing tests. */
-    val config = Config(
-        root = Paths.get("./cottontaildb-test"),
-        cli = false,
-        mapdb = MapDBConfig(enableMmap = false, false)
-    )
-
     /** General size of collections used for testing. */
-    const val collectionSize: Int = 100_000
+    const val collectionSize: Int = 50_000
 
     /** Maximum dimension used for vector generation. */
     const val smallVectorMaxDimension: Int = 128
@@ -33,4 +22,14 @@ object TestConstants {
 
     /** Maximum dimension used for vector generation. */
     const val largeVectorMaxDimension: Int = 2048
+
+    /**
+     * Creates a new test configuration.
+     *
+     * @return [Config]
+     */
+    fun testConfig() = Config(
+        root = Files.createTempDirectory("cottontaildb-test-${UUID.randomUUID()}"),
+        cli = false
+    )
 }